MNT Remove old code
