+++ README.md
@@ -46,10 +46,6 @@ use SilverStripe\LinkField\Form\LinkField;
 
 class Page extends SiteTree
 {
-    private static array $db = [
-        'DbLink' => DBLink::class
-    ];
-
     private static array $has_one = [
         'HasOneLink' => Link::class,
     ];
@@ -63,8 +59,8 @@ class Page extends SiteTree
             [
                 LinkField::create('HasOneLink'),
                 LinkField::create('DbLink'),
-            ]
-        )
+            ],
+        );
 
         return $fields;
     }
