--- src/JsonData.php
@@ -1,17 +1,0 @@
-<?php declare(strict_types=1);
-
-namespace SilverStripe\LinkField;
-
-use JsonSerializable;
-
-/**
- * An object that can be serialized and deserialized to JSON.
- */
-interface JsonData extends JsonSerializable
-{
-    /**
-     * @param array|JsonData $data
-     * @return $this
-     */
-    public function setData($data): self;
-}
