--- templates/SilverStripe/LinkField/Form/JsonField.ss
@@ -1,2 +1,0 @@
-<input $AttributesHTML />
-<div data-field-id="$ID" data-schema-component="$SchemaComponent" class="entwine-jsonfield"></div>
