+++ templates/SilverStripe/LinkField/Form/LinkField.ss
@@ -1,0 +1,2 @@
+<input $AttributesHTML />
+<div data-field-id="$ID" data-schema-component="$SchemaComponent" class="entwine-linkfield"></div>
